+++ pypump/pypump.py
@@ -350,8 +350,8 @@ class PyPump(object):
             response = fnc(url, **kwargs)
             return response
         except requests.exceptions.ConnectionError:
-            if self.protocol == "http" or raw:
-                raise  # shoot this seems real.
+            if (self.verify_requests and self.protocol == "https") or raw:
+                raise
             else:
                 self.set_http()
                 url = self._build_url(endpoint)
@@ -385,15 +385,19 @@ class PyPump(object):
 
     def construct_oauth_url(self):
         """ Constructs verifier OAuth URL """
-        response = requests.head("http://{0}".format(self.client.server))
+        response = requests.head("{0}://{1}".format(self.protocol, self.client.server))
         if response.is_redirect:
             server = response.headers['location']
+        elif self.protocol == "https" and not self.verify_requests:
+            self.set_http()
+            return self.construct_oauth_url()
         else:
             server = response.url
 
         path = "oauth/authorize?oauth_token={token}".format(
             token=self.store["oauth-request-token"]
         )
+        self.set_https()
         return "{server}{path}".format(
             server=server,
             path=path